update test_govern.py move govern_util.py to utils.platon_lib
