+++ utils/platon_lib/govern_util.py
@@ -2,9 +2,9 @@
 
 """
 @Author: 
-@Date: 2019/8/1 14:46
+@Date: block_interval19/8/1 14:46
 @LastEditors: 
-@LastEditTime: 2019/8/1 14:46
+@LastEditTime: block_interval19/8/1 14:46
 @Description:
 """
 import random
@@ -14,7 +14,6 @@ import json
 import os
 
 from client_sdk_python import Web3
-
 from common import log
 
 # 查询块高的时间间隔-单位秒
@@ -22,6 +21,9 @@ time_interval=10
 
 # 一个共识周期数包含的区块数
 # block_count
+
+# 截止块高在某个共识周期的第(block_count-block_interval)个块高
+# block_interval
 
 # 提案截止块高中，设置截止块高在第几个共识周期中
 # conse_index
@@ -70,10 +72,10 @@ def is_cur_block_number_big_than_end_block_number(rpc_link,end_number):
             break
 
 
-def get_single_valid_end_and_effect_block_number(rpc_link,block_count,conse_index):
+def get_single_valid_end_and_effect_block_number(rpc_link,block_count,block_interval,conse_index):
     '''
     获取构造单个合理的截止块高和生效块高
-    :param rpc_link,block_count,conse_index:
+    :param rpc_link,block_count,block_interval,conse_index:
     :return: tuple
     '''
     # 当前块高
@@ -82,25 +84,25 @@ def get_single_valid_end_and_effect_block_number(rpc_link,block_count,conse_inde
 
     if block_number % block_count == 0:
         # 截止块高
-        end_number = block_number + (conse_index * block_count) - 20
+        end_number = block_number + (conse_index * block_count) - block_interval
 
         # 生效块高
-        effect_number = end_number + (conse_index + 5) * block_count + 20
+        effect_number = end_number + (conse_index + 5) * block_count + block_interval
     else:
         mod = block_number % block_count
         interval = block_count - mod
         # 截止块高
-        end_number = block_number + interval + (conse_index * block_count) - 20
+        end_number = block_number + interval + (conse_index * block_count) - block_interval
 
         # 生效块高
-        effect_number = end_number + (conse_index + 5) * block_count + 20
+        effect_number = end_number + (conse_index + 5) * block_count + block_interval
     return end_number, effect_number
 
 
-def get_all_invalid_end_block_number(rpc_link,block_count,conse_index,conse_border):
+def get_all_invalid_end_block_number(rpc_link,block_count,block_interval,conse_index,conse_border):
     '''
     获取构造各类不合理的截止区块块高
-    :param rpc_link,block_count,conse_index,conse_border:
+    :param rpc_link,block_count,block_interval,conse_index,conse_border:
     :return: list
     '''
     # 当前块高
@@ -110,87 +112,89 @@ def get_all_invalid_end_block_number(rpc_link,block_count,conse_index,conse_bord
     if block_number % block_count == 0:
 
         # 截止块高
-        end_number = block_number + (conse_index * block_count) - 20
+        end_number = block_number + (conse_index * block_count) - block_interval
 
         # 生效块高
-        effect_number = end_number + (conse_index + 5) * block_count + 20
+        effect_number = end_number + (conse_index + 5) * block_count + block_interval
 
         end_number_list = [
             #  (None, effect_number),
             # ('number', effect_number),
             # ('0.a.0', effect_number),
             (block_number, effect_number),
-            (block_number + block_count * conse_index - 19, effect_number),
-            (block_number + block_count * conse_border - 21, effect_number + block_count * conse_border),
-            (block_number + block_count * (conse_border + 1) - 20, effect_number + block_count * (conse_border + 1))]
+            (block_number + block_count * conse_index - (block_interval-1), effect_number),
+            (block_number + block_count * conse_border - (block_interval+1), effect_number + block_count * conse_border),
+            (block_number + block_count * (conse_border + 1) - block_interval, effect_number + block_count * (conse_border + 1))]
     else:
         mod = block_number % block_count
         interval = block_count - mod
 
         # 截止块高
-        end_number = block_number + interval + (conse_index * block_count) - 20
+        end_number = block_number + interval + (conse_index * block_count) - block_interval
 
         # 生效块高
-        effect_number = end_number + (conse_index + 5) * block_count + 20
+        effect_number = end_number + (conse_index + 5) * block_count + block_interval
 
         end_number_list = [
             #  (None, effect_number),
             # ('number', effect_number),
             # ('0.a.0', effect_number),
             (block_number, effect_number),
-            (block_number + interval + block_count * conse_index - 19, effect_number),
-            (block_number + interval + block_count * conse_border - 21, effect_number + block_count * conse_border),
-            (block_number + interval + block_count * (conse_border + 1) - 20, effect_number + block_count * (conse_border + 1))]
+            (block_number + interval + block_count * conse_index - (block_interval-1), effect_number),
+            (block_number + interval + block_count * conse_border - (block_interval+1), effect_number + block_count * conse_border),
+            (block_number + interval + block_count * (conse_border + 1) - block_interval, effect_number + block_count * (conse_border + 1))]
     return end_number_list
 
 
-def get_all_invalid_effect_block_number(rpc_link,block_count,conse_index):
+def get_all_invalid_effect_block_number(rpc_link,block_count,block_interval,conse_index):
     '''
     获取构造各类不合理的生效区块块高
-    :param rpc_link,block_count,conse_index:
+    :param rpc_link,block_count,block_interval,conse_index:
     :return: list
     '''
     # 当前块高
     block_number = rpc_link.web3.eth.blockNumber
+
     log.info('当前块高={}'.format(block_number))
     
     if block_number % block_count == 0:
         # 截止块高
-        end_number = block_number + (conse_index * block_count) - 20
+        end_number = block_number + (conse_index * block_count) - block_interval
         effect_number_list = [
                               #   (end_number, None),
                               # (end_number, 'number'),
                               # (end_number, '0.a.0'),
                               (end_number, block_number),
                               (end_number, end_number),
-                              (end_number, end_number + (conse_index + 4) * block_count+20),
-                              (end_number, end_number + (conse_index + 5) * block_count - 1+20),
-                              (end_number, end_number + (conse_index + 5) * block_count + 1+20),
-                              (end_number, end_number + (conse_index + 11) * block_count+20)]
+                              (end_number, end_number + (conse_index + 4) * block_count+block_interval),
+                              (end_number, end_number + (conse_index + 5) * block_count - 1+block_interval),
+                              (end_number, end_number + (conse_index + 5) * block_count + 1+block_interval),
+                              (end_number, end_number + (conse_index + 10) * block_count+block_interval)]
     else:
         mod = block_number % block_count
         interval = block_count - mod
+        log.info(interval)
 
         # 截止块高
-        end_number = block_number + interval + (conse_index * block_count) - 20
+        end_number = block_number + interval + (conse_index * block_count) - block_interval
         effect_number_list = [
                               #   (end_number, None),
                               # (end_number, 'number'),
                               # (end_number, '0.a.0'),
                               (end_number, block_number),
                               (end_number, end_number),
-                              (end_number, end_number + (conse_index + 4) * block_count+20),
-                              (end_number, end_number + (conse_index + 5) * block_count - 1+20),
-                              (end_number, end_number + (conse_index + 5) * block_count + 1+20),
-                              (end_number, end_number + (conse_index + 11) * block_count+20)]
+                              (end_number, end_number + (conse_index + 4) * block_count+block_interval),
+                              (end_number, end_number + (conse_index + 5) * block_count - 1+block_interval),
+                              (end_number, end_number + (conse_index + 5) * block_count + 1+block_interval),
+                              (end_number, end_number + (conse_index + 10) * block_count+block_interval)]
 
     return effect_number_list
 
 
-def get_all_legal_end_and_effect_block_number(rpc_link,block_count,conse_index,conse_border):
+def get_all_legal_end_and_effect_block_number(rpc_link,block_count,block_interval,conse_index,conse_border):
     '''
     获取构造各类合理的截止区块、生效区块块高
-    :param rpc_link,block_count,conse_index,conse_border:
+    :param rpc_link,block_count,block_interval,conse_index,conse_border:
     :return: list
     '''
     # 当前块高
@@ -198,28 +202,28 @@ def get_all_legal_end_and_effect_block_number(rpc_link,block_count,conse_index,c
     log.info('当前块高={}'.format(block_number))
 
     if block_number % block_count == 0:
-        end_number_list = [(block_number + block_count * conse_index - 20,
+        end_number_list = [(block_number + block_count * conse_index - block_interval,
                             block_number + block_count * conse_index + (conse_index + 5) * block_count),
-                           (block_number + block_count * conse_index - 20,
+                           (block_number + block_count * conse_index - block_interval,
                             block_number + block_count * conse_index + (conse_index + 8) * block_count),
-                           (block_number + block_count * conse_border - 20,
+                           (block_number + block_count * conse_border - block_interval,
                             block_number + block_count * conse_border + (conse_border + 10) * block_count)]
     else:
         mod = block_number % block_count
         interval = block_count - mod
-        end_number_list = [(block_number + interval + block_count * conse_index - 20,
+        end_number_list = [(block_number + interval + block_count * conse_index - block_interval,
                             block_number + interval + block_count * conse_index + (conse_index + 5) * block_count),
-                           (block_number + interval + block_count * conse_index - 20,
+                           (block_number + interval + block_count * conse_index - block_interval,
                             block_number + interval + block_count * conse_index + (conse_index + 8) * block_count),
-                           (block_number + interval + block_count * conse_border - 20,
+                           (block_number + interval + block_count * conse_border - block_interval,
                             block_number + interval + block_count * conse_border + (conse_border + 10) * block_count)]
     return end_number_list
 
 
-def get_all_legal_end_and_effect_block_number_for_vote(rpc_link,block_count,conse_index,conse_border):
+def get_all_legal_end_and_effect_block_number_for_vote(rpc_link,block_count,block_interval,conse_index,conse_border):
     '''
     获取构造各类合理的截止区块、生效区块块高
-    :param rpc_link,block_count,conse_index,conse_border:
+    :param rpc_link,block_count,block_interval,conse_index,conse_border:
     :return: list
     '''
     # 当前块高
@@ -228,18 +232,18 @@ def get_all_legal_end_and_effect_block_number_for_vote(rpc_link,block_count,cons
 
     if block_number % block_count == 0:
         end_number_list = [
-            (block_number + block_count * conse_index - 20, block_number + block_count * conse_index + (conse_index + 5) * block_count),
-            (block_number + block_count * (conse_border - 1) - 20,
+            (block_number + block_count * conse_index - block_interval, block_number + block_count * conse_index + (conse_index + 5) * block_count),
+            (block_number + block_count * (conse_border - 1) - block_interval,
              block_number + block_count * (conse_border - 1) + (conse_border - 1 + 5) * block_count),
-            (block_number + block_count * conse_border - 20, block_number + block_count * conse_border + (conse_border + 5) * block_count)]
+            (block_number + block_count * conse_border - block_interval, block_number + block_count * conse_border + (conse_border + 5) * block_count)]
     else:
         mod = block_number % block_count
         interval = block_count - mod
-        end_number_list = [(block_number + interval + block_count * conse_index - 20,
+        end_number_list = [(block_number + interval + block_count * conse_index - block_interval,
                             block_number + interval + block_count * conse_index + (conse_index + 5) * block_count),
-                           (block_number + interval + block_count * (conse_border - 1) - 20,
+                           (block_number + interval + block_count * (conse_border - 1) - block_interval,
                             block_number + interval + block_count * (conse_border - 1) + (conse_border - 1 + 5) * block_count),
-                           (block_number + interval + block_count * conse_border - 20,
+                           (block_number + interval + block_count * conse_border - block_interval,
                             block_number + interval + block_count * conse_border + (conse_border + 5) * block_count)]
     return end_number_list
 
@@ -301,7 +305,7 @@ def waite_to_settle_account_cycle_block_number(rpc_link, block_count,end_block=N
 
     if not end_block:
         wait_block = block_count - (current_block % block_count)
-    end_block = current_block + wait_block
+        end_block = current_block + wait_block
 
     while 1:
         current_block = rpc_link.eth.blockNumber
@@ -313,42 +317,114 @@ def waite_to_settle_account_cycle_block_number(rpc_link, block_count,end_block=N
 
 def is_exist_ineffective_proposal_info(rpc_link):
     '''
-    判断链上是否存在有效的升级提案
+    判断链上是否存在有效的升级提案-用于判断是否可以发起升级提案
     :param rpc_link:
     :return:
     '''
+    log.info('is_exist_ineffective_proposal_info-开始')
     result = rpc_link.listProposal()
-    print(result)
-    proposalinfo = result.get('Data')
-    proposalinfo = json.loads(proposalinfo)
-    if proposalinfo == 'null':
-        return False
+
+    proposal_info = result.get('Data')
+
+    if not proposal_info:
+        log.info('查询提案失败')
+        log.info('is_exist_ineffective_proposal_info-结束')
     else:
-        proposal_list = []
-        endvotingblock_list = []
-        activeblock_list = []
-        # 查询当前块高
-        block_number = rpc_link.eth.blockNumber
+        log.info('查询提案成功')
+        proposal_info = json.loads(proposal_info)
+        if proposal_info is None:
+            log.info('提案信息为空')
+            log.info('is_exist_ineffective_proposal_info-结束')
+            return False
+        else:
+            log.info('有提案信息')
 
-        for p_list in proposalinfo:
-            proposal_list.append(p_list.get('ProposalID'))
-            endvotingblock_list.append(p_list.get('EndVotingBlock'))
-            activeblock_list.append(p_list.get('ActiveBlock'))
+            proposal_list = []
+            endvotingblock_list = []
 
-        flag = False
-        # 遍历升级提案
-        print(proposal_list)
-        for i in range(0, len(proposal_list)):
-            result = rpc_link.getTallyResult(proposal_list[i].strip())
-            if not result.get('Data'):
-                log.info('根据升级提案号，查询升级提案失败')
-            else:
-                status_info = result.get('Data')
-                status_info = json.loads(status_info)
-                status = status_info.get('status')
-                if status == 4 or status == 0:
+            for p_list in proposal_info:
+                proposal_list.append(p_list.get('ProposalID'))
+                endvotingblock_list.append(p_list.get('EndVotingBlock'))
+
+            log.info('提案信息为：{}'.format(proposal_info))
+            flag=False
+
+            #当前块高
+            block_number = rpc_link.eth.blockNumber
+            for i in range(0, len(proposal_list)):
+                if endvotingblock_list[i] > block_number:
+                    log.info('有可投票的升级提案')
                     flag = True
-        return flag
+                    break
+                else:
+                    result = rpc_link.getTallyResult(proposal_list[i].strip())
+                    data=result.get('Data')
+
+                    if not data:
+                        log.info('根据升级提案号，查询升级提案失败')
+                        flag = False
+                    else:
+                        status_info = json.loads(data)
+                        status = status_info.get('status')
+                        log.info('判断提案信息的状态={}'.format(status))
+                        if status == 4:
+                            log.info('有预生效的升级提案')
+                            flag = True
+                            break
+                        else:
+                            flag=False
+                            log.info('没有预生效的升级提案')
+            log.info('is_exist_ineffective_proposal_info-结束')
+            return flag
+
+
+def is_exist_ineffective_proposal_info_for_vote(rpc_link):
+    '''
+    判断链上是否存在有效的升级提案-用于判断是否可以发起投票
+    :param rpc_link:
+    :return:
+    '''
+    log.info('is_exist_ineffective_proposal_info_for_vote-开始')
+
+    result = rpc_link.listProposal()
+    proposal_info = result.get('Data')
+
+    if not proposal_info:
+        log.info('查询提案失败')
+        log.info('is_exist_ineffective_proposal_info_for_vote-结束')
+    else:
+        log.info('查询提案成功')
+        proposal_info = json.loads(proposal_info)
+
+        if proposal_info is None:
+            log.info('提案信息为空')
+            log.info('is_exist_ineffective_proposal_info_for_vote-结束')
+            return False
+        else:
+            log.info('有提案信息')
+            proposal_list = []
+            endvotingblock_list = []
+
+            for p_list in proposal_info:
+                proposal_list.append(p_list.get('ProposalID'))
+                endvotingblock_list.append(p_list.get('EndVotingBlock'))
+
+            log.info('提案信息为：{}'.format(proposal_info))
+
+            flag = False
+
+            #当前块高
+            block_number = rpc_link.eth.blockNumber
+            for i in range(0, len(proposal_list)):
+                if endvotingblock_list[i] > block_number:
+                    log.info('有可投票的升级提案')
+                    flag = True
+                    break
+                else:
+                    flag=False
+                    log.info('没有可投票的升级提案')
+            log.info('is_exist_ineffective_proposal_info_for_vote-结束')
+            return flag
 
 
 def get_effect_proposal_id(rpc_link):
@@ -357,18 +433,68 @@ def get_effect_proposal_id(rpc_link):
     :param rpc_link:
     :return:
     '''
-    result = rpc_link.listProposal()
-    proposal_info = result.get('Data')
-
     if not is_exist_ineffective_proposal_info(rpc_link):
         log.info('链上不存在有效的升级提案')
+        return None
     else:
+        result = rpc_link.listProposal()
+        proposal_info = result.get('Data')
         proposal_info = json.loads(proposal_info)
+
         proposalid_list = []
+        newversion_list = []
+        endvotingblock_list = []
 
         for pid_list in proposal_info:
             proposalid_list.append(pid_list.get('ProposalID'))
-        return proposalid_list[0]
+            endvotingblock_list.append(pid_list.get('EndVotingBlock'))
+            newversion_list.append(pid_list.get('NewVersion'))
+
+        # 当前块高
+        block_number = rpc_link.eth.blockNumber
+        for i in range(0, len(proposalid_list)):
+            if endvotingblock_list[i] > block_number:
+                return proposalid_list[i]
+            else:
+                result = rpc_link.getTallyResult(proposalid_list[i].strip())
+                if not result.get('Data'):
+                    log.info('根据升级提案号，查询升级提案失败')
+                else:
+                    status_info = result.get('Data')
+                    status_info = json.loads(status_info)
+                    status = status_info.get('status')
+                    if status == 4:
+                        return proposalid_list[i]
+
+
+def get_effect_proposal_info_for_vote(rpc_link):
+    '''
+    获取链上有效的升级提案版本号（暂时只支持升级提案）
+    :param rpc_link:
+    :return:
+    '''
+    if not is_exist_ineffective_proposal_info_for_vote(rpc_link):
+        log.info('链上不存在可投票的升级提案')
+        return None
+    else:
+        result = rpc_link.listProposal()
+        proposal_info = result.get('Data')
+        proposal_info = json.loads(proposal_info)
+
+        proposalid_list = []
+        newversion_list = []
+        endvotingblock_list = []
+
+        for pid_list in proposal_info:
+            proposalid_list.append(pid_list.get('ProposalID'))
+            newversion_list.append(pid_list.get('NewVersion'))
+            endvotingblock_list.append(pid_list.get('EndVotingBlock'))
+
+        # 当前块高
+        block_number = rpc_link.eth.blockNumber
+        for i in range(0, len(proposalid_list)):
+            if endvotingblock_list[i] > block_number:
+                return proposalid_list[i],newversion_list[i],endvotingblock_list[i]
 
 
 def get_proposal_vote_end_block_number(rpc_link):
@@ -388,9 +514,8 @@ def get_proposal_vote_end_block_number(rpc_link):
 
         for e_list in proposal_info:
             end_block_list.append(e_list.get('EndVotingBlock'))
-        end_block_list = end_block_list.sort()
-        print(end_block_list)
-        return end_block_list[len(end_block_list)-1]
+        end_block_list.sort()
+        return end_block_list[-1]
 
 
 def get_proposal_vote_effect_block_number(rpc_link):
@@ -410,35 +535,39 @@ def get_proposal_vote_effect_block_number(rpc_link):
 
         for e_list in proposal_info:
             effect_block_list.append(e_list.get('ActiveBlock'))
-        effect_block_list = effect_block_list.sort()
+        effect_block_list.sort()
 
-        return effect_block_list[len(effect_block_list)-1]
+        return effect_block_list[-1]
 
 
 def get_effect_proposal_id_and_version(rpc_link):
     '''
-    获取当前提案列表，第一个提案的提案 ProposalID NewVersion（暂时只支持升级提案，所有获取到的是当前生效的升级提案id）
+    获取有效的升级提案的 ProposalID NewVersion（暂时只支持升级提案，所有获取到的是当前生效的升级提案 ProposalID NewVersion）
     :param rpc_link:
     :return:
     '''
-    result = rpc_link.listProposal()
-    proposal_info = result.get('Data')
-    proposal_list = []
-
     if not is_exist_ineffective_proposal_info(rpc_link):
         log.info('链上不存在有效的升级提案')
-        proposal_list.append('proposal')
+        return None
     else:
-        log.info('链上存在有效的升级提案')
+        result = rpc_link.listProposal()
+        proposal_info = result.get('Data')
         proposal_info = json.loads(proposal_info)
 
-        proposal_id=proposal_info[0].get('ProposalID')
-        new_version=proposal_info[0].get('NewVersion')
+        proposalid_list = []
+        newversion_list = []
+        endvotingblock_list =[]
 
-        proposal_list.append(proposal_id)
-        proposal_list.append(new_version)
+        for pid_list in proposal_info:
+            proposalid_list.append(pid_list.get('ProposalID'))
+            endvotingblock_list.append(pid_list.get('EndVotingBlock'))
+            newversion_list.append(pid_list.get('NewVersion'))
 
-    return proposal_list
+        #当前块高
+        block_number = rpc_link.eth.blockNumber
+        for i in range(0, len(proposalid_list)):
+            if endvotingblock_list[i] > block_number:
+                return proposalid_list[i], newversion_list[i]
 
 
 def get_version(rpc_link,flag=None):
@@ -487,6 +616,91 @@ def get_version(rpc_link,flag=None):
     else:
         pass
     return new_version
+
+
+def get_rate_of_voting(rpc_link, proposalid):
+    '''
+    计算升级提案投票率
+    :param rpc_link:
+    :param proposalid:
+    :return:
+    '''
+    result = rpc_link.getTallyResult(proposalid)
+    resultinfo = result.get('Data')
+    resultinfo = json.loads(resultinfo)
+
+    if not resultinfo:
+        log.info('根据给定的提案id查询提案结果失败')
+    else:
+        yeas = resultinfo.get('yeas')
+        accuVerifiers = resultinfo.get('accuVerifiers')
+        return yeas/accuVerifiers
+
+
+def get_accuVerifiers_of_proposal(rpc_link, proposalid):
+    '''
+    获取在整个投票期内有投票资格的验证人总数
+    :param rpc_link:
+    :param proposalid:
+    :return:
+    '''
+    result = rpc_link.getTallyResult(proposalid)
+    resultinfo = result.get('Data')
+    resultinfo = json.loads(resultinfo)
+
+    if not resultinfo:
+        log.info('根据给定的提案id查询提案结果失败')
+    else:
+        accuVerifiers = resultinfo.get('accuVerifiers')
+        return accuVerifiers
+
+
+def get_yeas_of_proposal(rpc_link, proposalid):
+    '''
+    获取在整个投票期内有投票资格的验证人总数
+    :param rpc_link:
+    :param proposalid:
+    :return:
+    '''
+    result = rpc_link.getTallyResult(proposalid)
+    resultinfo = result.get('Data')
+    resultinfo = json.loads(resultinfo)
+
+    if not resultinfo:
+        log.info('根据给定的提案id查询提案结果失败')
+    else:
+        yeas = resultinfo.get('yeas')
+        return yeas
+
+
+def get_current_verifier(rpc_link):
+    '''
+    获取当前结算周期验证人列表
+    :param rpc_link:
+    :return:
+    '''
+    result = rpc_link.getVerifierList()
+    verifier_info = result.get('Data')
+
+    verifier_list = []
+    for list1 in verifier_info:
+        verifier_list.append(list1.get('NodeId'))
+    return verifier_list
+
+
+def get_current_validator(rpc_link):
+    '''
+    获取当前共识轮验证人列表
+    :param rpc_link:
+    :return:
+    '''
+    result = rpc_link.getValidatorList()
+    validator_info = result.get('Data')
+
+    validator_list = []
+    for list1 in validator_info:
+        validator_list.append(list1.get('NodeId'))
+    return validator_list
 
 
 def gen_random_string(length):
